NEW UPDATE v5.4.0 - HOVER TO SEE
We’ve fixed the path files for commands. Instead of use pocketmine\event\TranslationContainer, it’s now use \pocketmine\lang\TranslationContainer. This is due to a new PMMP Update, which moved the TranslationContainer to another folder.
We’ve finally bumped the API to support ALPHA12.
We’ve bumped the plugin version to v5.4.0.
We’ve fixed issues where sometimes worldedit wouldn’t work, and you’d need to log off and log back on for it to fix. A weird bug which was fixed.
Fixed Commands issues + Bug fixes.